requirements.txt
playwright==1.52.0
pyyaml==6.0.3
